linuxspi: use linux/gpio instead of /sys accessors
/sys/class/gpio accessors are deprecated and ugly: They don't track the
state. If avrdude crashes, the GPIO won't recover.

Better use linux/gpio.h: It automatically recovers initial state when we
close the device.

The format of the port specification changed now, as it needs to know
which gpiochip device to use. The format is as follows:

-P /dev/spidev:/dev/gpiochip[:resetpin]

The reset pin at the end is optional. If not specified, the default
reset pin of avrdude's configuration will be used.

Signed-off-by: Ralf Ramsauer <ralf@vmexit.de>